+++ tina-runtime/src/lib.rs
@@ -61,6 +61,7 @@ mod drain_state;
 mod driver;
 mod errors;
 pub mod event_sink;
+mod fact;
 mod full_handling;
 pub mod guarded_pending;
 mod host_burst;
@@ -80,6 +81,7 @@ pub mod scope;
 pub mod service_pressure;
 pub mod sharded;
 pub mod shared_scope;
+pub mod shared_work;
 mod shutdown;
 mod single_call_gate;
 pub mod tcp_loops;
@@ -285,6 +287,11 @@ use driver::DriverCompletion;
 pub use event_sink::{
     BoundedEventSink, DropPolicy, EventSinkDrain, EventSinkReport, EventSinkSurface,
 };
+pub use fact::{
+    GrpcStatusCode, GrpcStreamId, Http2CloseReason, Http2FlowControlSide, Http2ResetReason,
+    Http2StreamId, IntoRuntimeFact, ProtocolConnectionId, ProtocolDirection, ProtocolFact,
+    ProtocolFamily, RuntimeFact, WebSocketCloseReason, WebSocketSessionId,
+};
 pub use guarded_pending::{
     GuardedInsertError, GuardedParkCallError, GuardedParkError, GuardedParkTicket,
     GuardedPendingReplies, GuardedReplyError, GuardedTakeError,
@@ -304,6 +311,10 @@ pub use scope::{
 };
 pub use service_pressure::{ServicePressureReport, ServicePressureSurface, ServiceSurfaceState};
 pub use shared_scope::{SharedCapacityScope, SharedLease, SharedScopeFull, SharedScopeReport};
+pub use shared_work::{
+    SharedWork, SharedWorkCallError, SharedWorkError, SharedWorkReplyError, SharedWorkSnapshot,
+    SharedWorkTicket, request_effect_after_shared_wait,
+};
 pub use tcp_loops::{LoopStep, ReadExactStep, TcpReadExact, TcpReadToEof, TcpWriteAll};
 /// Declares a Tina isolate whose call channel defaults to [`RuntimeCall<Message>`](RuntimeCall).
 ///
@@ -975,6 +986,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
         M: Mailbox<I::Message> + 'static,
     {
@@ -1004,6 +1016,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let address = self.register_entry::<I, Outbound>(
@@ -1083,6 +1096,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let address = self.register_with_capacity::<I, Outbound>(isolate, mailbox_capacity);
@@ -1107,6 +1121,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let address = self.register_with_capacity::<I, Outbound>(isolate, mailbox_capacity);
@@ -1141,6 +1156,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let address = self.register_with_capacity::<I, Outbound>(isolate, mailbox_capacity);
@@ -1174,6 +1190,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let mailbox = self
@@ -1233,6 +1250,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
         Ctor: FnOnce(Address<I::Message, I::Reply>) -> I,
     {
@@ -1997,6 +2015,14 @@ where
             }
             ErasedEffect::ReplyTo { handle, message } => {
                 self.execute_reply_to(isolate_id, cause, handle, message, route_remote);
+                false
+            }
+            ErasedEffect::Fact(fact) => {
+                self.push_event(
+                    isolate_id,
+                    Some(cause),
+                    RuntimeEventKind::FactObserved { fact },
+                );
                 false
             }
         }
@@ -3844,6 +3870,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         let isolate_id = IsolateId::new(self.next_isolate_id);
@@ -3883,6 +3910,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: Send + 'static,
     {
         let address = self.register_sendable_entry::<I, Outbound>(
@@ -3912,6 +3940,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: Send + 'static,
     {
         let mailbox = self
@@ -3950,6 +3979,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: Send + 'static,
     {
         let isolate_id = IsolateId::new(self.next_isolate_id);
@@ -3991,6 +4021,7 @@ where
         I::Spawn: IntoErasedSpawn<S, F> + 'static,
         I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
         I::Call: IntoErasedCall<I::Message> + 'static,
+        I::Fact: IntoRuntimeFact + 'static,
         Outbound: 'static,
     {
         if mailbox_capacity == 0 {
@@ -4296,6 +4327,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -4367,6 +4399,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: Send + 'static,
     S: Shard,
     F: MailboxFactory,
@@ -4430,6 +4463,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -4465,6 +4499,7 @@ where
             handle: tina::runtime_internal::deferred_into_handle(slot),
             message: ErasedMessage::Local(Box::new(reply)),
         },
+        Effect::Fact(fact) => ErasedEffect::Fact(fact.into_runtime_fact()),
     }
 }
 
@@ -4476,6 +4511,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: Send + 'static,
     S: Shard,
     F: MailboxFactory,
@@ -4511,6 +4547,7 @@ where
             handle: tina::runtime_internal::deferred_into_handle(slot),
             message: ErasedMessage::Sendable(Box::new(reply)),
         },
+        Effect::Fact(fact) => ErasedEffect::Fact(fact.into_runtime_fact()),
     }
 }
 
@@ -4550,6 +4587,7 @@ where
         handle: DeferredReplyHandle,
         message: ErasedMessage,
     },
+    Fact(RuntimeFact),
 }
 
 impl<S, F> ErasedEffect<S, F>
@@ -4571,6 +4609,7 @@ where
             Self::Call(_) => EffectKind::Call,
             Self::Batch(_) => EffectKind::Batch,
             Self::ReplyTo { .. } => EffectKind::ReplyTo,
+            Self::Fact(_) => EffectKind::Fact,
         }
     }
 
@@ -4932,6 +4971,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -4969,6 +5009,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     OutboundMsg: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -5002,6 +5043,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -5045,6 +5087,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     Outbound: 'static,
     S: Shard,
     F: MailboxFactory,
@@ -5069,6 +5112,7 @@ where
     I::Spawn: IntoErasedSpawn<S, F> + 'static,
     I::SpawnObserved: IntoErasedSpawnObserved<S, F, I::Message> + 'static,
     I::Call: IntoErasedCall<I::Message> + 'static,
+    I::Fact: IntoRuntimeFact + 'static,
     OutboundMsg: 'static,
     S: Shard,
     F: MailboxFactory,